requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
easyocr==1.7.0
opencv-python==4.8.1.78
Pillow==9.5.0
requests==2.31.0
gspread==5.11.3
oauth2client==4.1.3
pyzbar==0.1.9
numpy==1.24.3
langdetect==1.0.9
aiofiles==23.2.0